src/synth_smiles/requirement_full.txt
absl-py==2.3.0
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aizynthfinder==4.3.2
alabaster==1.0.0
alembic==1.17.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
apted==1.0.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
autopage==0.5.2
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.3.0
blinker==1.9.0
blosc2==3.7.2
botorch==0.14.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
chemprop==1.5.2
click==8.2.1
cliff==4.12.0
cloudpickle==3.1.1
cmaes==0.12.0
cmd2==2.7.0
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.0
docopt==0.6.2
docstring_parser==0.16
docutils==0.21.2
entrypoints==0.4
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.9.23
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
fuzzywuzzy==0.18.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gpytorch==1.14
greenlet==3.2.4
GridDataFormats==1.0.2
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
hyperopt==0.2.7
idna==3.10
imagesize==1.4.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.37.0
ipython-genutils==0.2.0
ipywidgets==7.8.5
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.3.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==7.4.9
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==1.1.11
jupytext==1.18.1
kiwisolver==1.4.8
lark==1.3.1
lightning-utilities==0.15.2
linear-operator==0.6
llvmlite==0.45.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.2.1
mda-xdrlib==0.2.0
MDAnalysis==2.9.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.1.4
mmtf-python==1.1.3
molbloom==2.3.5
mpmath==1.3.0
mrcfile==1.5.4
msgpack==1.1.2
multidict==6.4.4
multipledispatch==1.0.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nbclassic==1.3.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.1
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.9.2
notebook==6.5.7
notebook_shim==0.2.4
numba==0.62.1
numexpr==2.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.16.3
openbabel-wheel==3.1.1.22
opt_einsum==3.4.0
optuna==2.10.1
overrides==7.7.0
packaging==25.0
pandas==1.3.5
pandas_flavor==0.7.0
pandocfilters==1.5.1
paretoset==1.2.5
parso==0.8.5
pathos==0.3.4
peft==0.17.0
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.8
polyleven==0.9.0
pox==0.3.6
ppft==1.7.7
prettytable==3.17.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==3.20.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py4j==0.10.9.9
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pyDeprecate==0.3.2
Pygments==2.19.2
pymongo==3.13.0
pyparsing==3.2.3
pyperclip==1.11.0
pyre-extensions==0.0.32
pyro-api==0.1.2
pyro-ppl==1.9.1
PySocks==1.7.1
PyTDC==0.4.0
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytorch-lightning==1.6.5.post0
pytorch-tree-lstm==0.1.3
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
rd_filters @ git+https://github.com/PatWalters/rd_filters.git@3630f1a59150b4a203c2f06d8376a66d1d4c3f70
rdchiral==1.1.0
rdkit==2023.9.6
rdkit-pypi==2022.9.5
referencing==0.37.0
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-argparse==1.7.2
route-distances==1.2.4
rpds-py==0.29.0
-e git+https://github.com/SeonghwanSeo/RxnFlow.git@23017e364017138cf56e11ebe5a171cfdc58aeef#egg=rxnflow
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.13.2
selfies==2.2.0
Send2Trash==1.8.3
sentry-sdk==2.31.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.44
stack-data==0.6.3
stevedore==5.6.0
sympy==1.13.1
tables==3.10.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
terminado==0.18.1
threadpoolctl==3.6.0
timeout-decorator==0.5.0
tinycss2==1.4.0
tokenizers==0.13.3
tomli==2.2.1
torch==2.5.1
torch-geometric==2.6.1
torch_cluster==1.6.3+pt25cu121
torch_scatter==2.1.2+pt25cu121
torch_sparse==0.6.18+pt25cu121
torchmetrics==1.8.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.32.1
triton==3.1.0
typed-argument-parser==1.10.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unidock_tools @ git+https://github.com/dptech-corp/Uni-Dock.git@6c4c72da22755b8dcb336c18b96ed62d01993c2f#subdirectory=unidock_tools
uri-template==1.3.0
urllib3==2.5.0
wadler_lindig==0.1.7
wandb==0.20.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==3.6.10
wrapt==2.0.1
xarray==2025.6.1
xformers==0.0.28.post3
yarl==1.20.1
